Make change trigger deferrable and optionally immediately deferred (#61)
This patch makes the change trigger `DEFERRABLE` by changing the
behaviour of `Carbonite.Migrations.create_trigger/2`. Existing triggers
will not be changed.

A new `:initially` option can be used to set the constraint to
`INITIALLY DEFERRED` and hence make it evaluate at the end of a
DB transaction. This allows to insert the `Carbonite.Transaction` only
conditionally *at the end of the transaction*.

Relates to [[#60]](https://github.com/bitcrowd/carbonite/issues/60)