Add :to option to Carbonite.override_mode/2 (#80)
This PR adds the `:to` option to `Carbonite.override_mode/2` which
allows to directly set a target mode on the trigger configuration (of
all tables, as before). This is useful when one wants to temporarily set
the mode to `:ignore` (e.g. during test setup).

Also improved the documentation on testing / bypassing Carbonite a bit.

Relates to #79