[#82] Array changes are not always detected (#83)
[[#82]](https://github.com/bitcrowd/carbonite/issues/82)

This patch fixes a bug in the trigger that prevented changes to array
fields from being detected correctly. Root cause was the use of the
"jsonb containment" operator to test whether a `(key, old value)` tuple
was "contained" within the JSON object of the updated record, where
instead we should have properly tested the values for equality.

Fixes #82.