[#86] Fix incompatibility with PostgreSQL 16.0
[[#86]](https://github.com/bitcrowd/carbonite/issues/86)

This patch fixes outdated migration V3 that set a column to default to
`0::TEXT::xid8` which became invalid syntax in Postgres 16. The relevant
column has since been updated to a `BIGINT` and the default is set to `0`
in migration V4.

Selected fix is to replace the invalid syntax in migration V3 with `0`
which seems to work fine with xid8 columns as well. Tested in Postgres
13.12 and 16.0.

Fixes #86.