Make set_transaction_id use the PK index for its lookup
This patch fixes a longstanding (yet undiscovered) bug in the `set_transaction_id`
procedure that caused it to not use the provided index on `id` spanning `xact_id`.
Reason was the use of `CURRVAL` in the `WHERE` clause which caused the query to
evaluate every row against a new function call, turning the should-be-fast lookup
into a seq scan of the `transactions` table. Consequently, `INSERT`s scaled linearly
with the size of the `transactions` table, instead of logarithmically.

Huge 🤦